Address design comments
