+++ src/components/Input/NumericInput/NumericInput.tsx
@@ -1,4 +1,5 @@
 import { useRef } from 'react';
+import { MdWarning } from 'react-icons/md';
 import { NumberInput, NumberInputProps, ThemeIcon, Tooltip } from '@mantine/core';
 
 import { usePropListeningState } from '@/hooks/util';
